models/krate: Inline `validate()` method
This function does not need to be on the `NewCrate` struct. Making it independent from the struct allows us to perform this validation much earlier in the process and combine it with other validations.
